+++ src/client/ServiceBusClient.ts
@@ -42,7 +42,22 @@ export default class ServiceBusClient implements IServiceBusClient {
     }
 
     public getSubscriptionDetails = async (topic: string, subscription: string): Promise<ISubscription> => {
-        return await this.getEntity<ISubscription>('GET', `${topic}/subscriptions/${subscription}`);
+        const subDetail = await this.getEntity<ISubscription>('GET', `${topic}/subscriptions/${subscription}`);
+        
+        // Strip out the d3p1: from CountDetails members
+        var leObject = subDetail.content.SubscriptionDescription.CountDetails as any;
+        Object.keys(subDetail.content.SubscriptionDescription.CountDetails).map(x=>{
+            leObject[x.substring(5)]= leObject[x];
+        });
+        //
+
+        return subDetail;
+    }
+
+    public createSubscription = async (topic: string, subscription: string): Promise<ISubscription> => {
+        const bodyContent = '<?xml version="1.0" encoding="utf-8" ?><entry xmlns="http://www.w3.org/2005/Atom"><content type="application/xml"><SubscriptionDescription xmlns="http://schemas.microsoft.com/netservices/2010/10/servicebus/connect"/></content></entry>';
+        const result = await this.sendRequest('PUT', `${topic}/subscriptions/${subscription}`, bodyContent);
+        return result.entry;
     }
 
     public getMessages = async (topic: string, subscription: string, searchArguments: string | null): Promise<SBC.ReceivedMessageInfo[]> => {
@@ -93,7 +108,7 @@ export default class ServiceBusClient implements IServiceBusClient {
                     return [result.feed.entry];
                 }
                 else {
-                    return result.feed;
+                    return [];
                 }
             }
             else {
@@ -104,15 +119,15 @@ export default class ServiceBusClient implements IServiceBusClient {
         return [];
     }
 
-    private async sendRequest(method: string, path: string): Promise<any> {
+    private async sendRequest(method: string, path: string, bodyContent: string | undefined = undefined): Promise<any> {
         const { sasToken, endpoint } = this.auth;
 
-        var result = await fetch(endpoint.replace('sb', 'https') + path, {
+        var result = await fetch(endpoint.replace('sb', 'https') + path + '?api-version=2015-01', {
             method: method || 'GET',
             headers: {
-                'Authorization': sasToken,
-                // 'api-version': '2015-01'
+                'Authorization': sasToken
             },
+            body: bodyContent,
         });
 
         if (!result.ok) {
@@ -144,4 +159,31 @@ export default class ServiceBusClient implements IServiceBusClient {
             return body;
         }
     }
+
+    
+    public async sendMessage(topic: string, body: any, contentType: string): Promise<void> {
+        let client;
+
+        try {
+            client = SBC.ServiceBusClient.createFromConnectionString(this.connectionString);
+            const topicClient = client.createTopicClient(topic);
+
+            const sender = topicClient.createSender();
+
+            sender.send({
+                body: body,
+                contentType: contentType
+            });
+            
+
+            await client.close();
+           
+        } catch{
+
+            if (client) {
+                await client.close();
+            }
+        }
+    }
+
 }