+++ src/commands/index.ts
@@ -7,8 +7,10 @@ import { TopicList } from "../topic/topicList";
 import { QueueList } from "../queue/queueList";
 import { Subscription } from "../topic/subscription";
 import { SubscriptionUI } from "../topic/SubscriptionUI";
+import { SendToBus } from "../messages/sendToBus";
+import { Topic } from "../topic/topic";
 
-export default function registerCommands(context: ExtensionContext, serviceBusProvider: ServiceBusProvider, nameSpace: NameSpace, subscriptionUI: SubscriptionUI) : IDisposable[] {
+export default function registerCommands(context: ExtensionContext, serviceBusProvider: ServiceBusProvider, nameSpace: NameSpace, subscriptionUI: SubscriptionUI, sendToBus: SendToBus): IDisposable[] {
 	return [
 		commands.registerCommand('serviceBusExplorer.refreshEntry', () => serviceBusProvider.reBuildTree()),
 
@@ -24,20 +26,40 @@ export default function registerCommands(context: ExtensionContext, serviceBusPr
 
 		commands.registerCommand('serviceBusExplorer.deleteEntry', (node: NameSpaceItem) => serviceBusProvider.deleteNamespace(node)),
 
-		commands.registerCommand('serviceBusExplorer.refreshTopicList', (node: TopicList) => window.showInformationMessage('Refresh Topic List not implemented!')),
+		commands.registerCommand('serviceBusExplorer.refreshTopicList', (node: TopicList) => serviceBusProvider.reBuildTree(node)),
 
-		commands.registerCommand('serviceBusExplorer.refreshQueueList', (node: QueueList) => window.showInformationMessage('Refresh Queue List not implemented!')),
+		commands.registerCommand('serviceBusExplorer.refreshQueueList', (node: QueueList) => serviceBusProvider.reBuildTree(node)),
 
 		commands.registerCommand('serviceBusExplorer.getSubscriptionMessages', async (node: Subscription) => await node.getSubscriptionMessages(context)),
+		
 		commands.registerCommand('serviceBusExplorer.searchMessage', async (node: Subscription) => {
-			var state  = await  subscriptionUI.searchMessages();
+			var state = await subscriptionUI.searchMessages();
 			await node.searchMessages(context, state.searchArguments);
 		}),
-		
+
 		commands.registerCommand('serviceBusExplorer.showMessage', async (topic: string, subscription: string, message: any) => {
 			let uri = Uri.parse(`servicebusmessage:message_${message.messageId}.json?topic=${topic}&subscription=${subscription}&messageid=${message.messageId}`);
 			let doc = await workspace.openTextDocument(uri); // calls back into the provider
 			await window.showTextDocument(doc, { preview: false });
+		}),
+
+		commands.registerCommand('serviceBusExplorer.sendToBus', async () => {
+			if (window.activeTextEditor) {
+				const documentText = window.activeTextEditor.document.getText();
+				var destination = await sendToBus.sendToBus();
+				await destination.client.sendMessage(destination.selectedTopic.title, documentText, 'application/json');
+				window.showInformationMessage('Message sent successfully');
+			}
+			else {
+				window.showErrorMessage('Only implemented for active document');
+			}
+
+		}),		
+		
+		commands.registerCommand('serviceBusExplorer.createSubscription', async (node: Topic) => {
+			var state  = await  subscriptionUI.createSubscription();
+			await node.createSubscription(context, state.name);
+			await serviceBusProvider.reBuildTree(node);
 		})
 	];
 }