+++ src/topic/topic.ts
@@ -1,9 +1,8 @@
-import { TreeItemCollapsibleState, Command } from "vscode";
+import { TreeItemCollapsibleState, ExtensionContext, Command, window } from "vscode";
 import { ExplorerItemBase, IItemData } from "../common/explorerItemBase";
 import { Subscription } from "./subscription";
 import { ISubscription } from "../client/models/ISubscriptionDetails";
 import path from 'path';
-
 
 export class Topic extends ExplorerItemBase {
 
@@ -51,6 +50,12 @@ export class Topic extends ExplorerItemBase {
 		}
 
 		return Promise.resolve([]);
+	}
+
+	public createSubscription = async (context: ExtensionContext, newSubscriptionName: string) => {
+		if (this.itemData.clientInstance) {
+			await this.itemData.clientInstance.createSubscription(this.label || '', newSubscriptionName);
+		}
 	}
 
 	contextValue = 'topic';